backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Xinlyu Wang <xinlalalele@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.0.7"
django-cors-headers = "4.3.1"
djangorestframework = "3.15.1"
gunicorn = "22.0.0"
packaging = "24.0"
psycopg2-binary = "2.9.9"
sqlparse = "0.5.0"
typing-extensions = "4.11.0"
djangocms-installer = "^2.0.0"
django-environ = "^0.11.2"
django-cms = "^4.1.3"
django-treebeard = "^4.7.1"
djangocms-admin-style = "^3.3.1"
django-filer = "^3.2.3"
djangocms-versioning = "^2.1.0"
djangocms-alias = "^2.0.1"
parler = "^1.0.1"
djangocms-text = "^0.3.2"
djangocms-frontend = "^1.3.4"
python-decouple = "^3.8"
cryptography = "^44.0.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
